Create first components
At the moment the project has the App.js and index.js files, but also
a new folder named "components", in which the files navbar.js,
mainBody.js, and title.js were created to tackle specific sections of
the website.

In the navbar component a form was created to input the desired
information to be displayed on the CV. At the moment only the user's
name can be introduced.