Add component for personal personalData
The personal data includes the user's career, cell phone number and
email address.

In the App.js file new states were added, which correspond to the
career, cell phone and email, which are passed in to the navbar
component to change their state. Likewise, these states are passed in to
the mainBody component, which takes care of updating the DIV for the
personal data.