Add skills component
The component is located on the left side of the CV document. At the
moment it contains the name of the skill, which can be introduced by
means of a new form and a single input.

The skill also has a bar to indicate the experience the user has with
that specific skill. Furthermore, it contains a button to erase the
skill.