Improve skill component
The component was removed from the navbar and was replaced with a button
that opens a new form to add a new Skill. This was done by moving the
form's logic from the navbar component to the skillsForm.js file, which
creates a new component exclusively for the skill's form.

Once the user submits a skill the new form is closed by changing its
display to "none".