Create Experience component
This component oversees the creation of the experience data entered by
the user. The props, which were defined in the App file, are passed in
and created by means of map. Map is apllied to any of the props to be
able to traverse across the array in order to create a new LI element
for each element.