Fix bug for removing skills
At the outset the skills were deleted using the filter function, however
the difficulty could be the same for other skills, which lead to the
removal of several difficulties, producing different bugs.

The problem was solved by using the slice function, in which an array
of the elements starting from 0 to the index of the element to be
removed, and then from the that element index + 1 to the end, was
assigned as the new corresponding state.