Add logic to edit any experience on the CV
This was done by using an SVG to trigger the event. In the SVG the ID
was defined with the help of the iterator used in the corresponding
map function. This ID is the same for both the SVG element and its
path, as any of those elements trigger the event.

The ID is used to access the information of each state to retrieve it,
so a new form is opened with the corresponding values in each input.
Once the form is submited, the states are updated to display the edited
information.